# data/uwbfuncpages.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "UWB 定位系统 - 泳池定位项目",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
    "/sysuser/listall": {
      "get": {
        "summary": "列举系统用户",
        "x-apifox-folder": "系统用户相关接口",
        "x-apifox-status": "testing",
        "deprecated": false,
        "description": "",
        "tags": [
          "系统用户相关接口"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": true,
            "example": "123",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-ignore-properties": [],
                  "x-apifox-orders": []
                }
              }
            }
          }
        },
        "x-run-in-apifox'": "https://www.apifox.cn/web/project/1888020/apis/api-48211717-run"
      }
    },
    "/sysuser": {
      "post": {
        "summary": "新增系统用户",
        "x-apifox-folder": "系统用户相关接口",
        "x-apifox-status": "testing",
        "deprecated": false,
        "description": "`SportsCompanyID` 和 `SportsCompanyName` 由 前端 填写，后端不会进行干预。\n`ModifyDatetime` 和 `CreateDatetime` 无论前端填写什么内容，后端都会干预。\n`Creator` 和 `Modifier` 若前端未填写，则后端会默认使用 `admin` 作为数据修改/新增者。\n",
        "tags": [
          "系统用户相关接口"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": false,
            "example": "123",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SysUser"
              },
              "example": {}
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-ignore-properties": [],
                  "x-apifox-orders": []
                }
              }
            }
          }
        },
        "x-run-in-apifox'": "https://www.apifox.cn/web/project/1888020/apis/api-48223918-run"
      },
      "get": {
        "summary": "通过登录名获取用户信息",
        "x-apifox-folder": "系统用户相关接口",
        "x-apifox-status": "testing",
        "deprecated": false,
        "description": "",
        "tags": [
          "系统用户相关接口"
        ],
        "parameters": [
          {
            "name": "login_name",
            "in": "query",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "StatusCode": {
                      "type": "integer"
                    },
                    "Message": {
                      "type": "object",
                      "properties": {
                        "Id": {
                          "type": "string"
                        },
                        "LoginName": {
                          "type": "string"
                        },
                        "DisplayName": {
                          "type": "string"
                        },
                        "Cellphone": {
                          "type": "string"
                        },
                        "Email": {
                          "type": "string"
                        },
                        "Wechat": {
                          "type": "string"
                        },
                        "QQ": {
                          "type": "string"
                        },
                        "IsSiteOwner": {
                          "type": "boolean"
                        },
                        "IsDisableLogin": {
                          "type": "boolean"
                        },
                        "CreateDatetime": {
                          "type": "string"
                        },
                        "ModifyDatetime": {
                          "type": "string"
                        },
                        "Creator": {
                          "type": "string"
                        },
                        "Modifier": {
                          "type": "string"
                        },
                        "PasswdMD5": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "Id",
                        "LoginName",
                        "DisplayName",
                        "Cellphone",
                        "Email",
                        "Wechat",
                        "QQ",
                        "IsSiteOwner",
                        "IsDisableLogin",
                        "CreateDatetime",
                        "ModifyDatetime",
                        "Creator",
                        "Modifier",
                        "PasswdMD5"
                      ],
                      "x-apifox-ignore-properties": [],
                      "x-apifox-orders": [
                        "Id",
                        "LoginName",
                        "DisplayName",
                        "Cellphone",
                        "Email",
                        "Wechat",
                        "QQ",
                        "IsSiteOwner",
                        "IsDisableLogin",
                        "CreateDatetime",
                        "ModifyDatetime",
                        "Creator",
                        "Modifier",
                        "PasswdMD5"
                      ]
                    }
                  },
                  "required": [
                    "StatusCode",
                    "Message"
                  ],
                  "x-apifox-ignore-properties": [],
                  "x-apifox-orders": [
                    "StatusCode",
                    "Message"
                  ]
                }
              }
            }
          }
        },
        "x-run-in-apifox'": "https://www.apifox.cn/web/project/1888020/apis/api-48227968-run"
      },
      "delete": {
        "summary": "删除系统用户",
        "x-apifox-folder": "系统用户相关接口",
        "x-apifox-status": "developing",
        "deprecated": false,
        "description": "",
        "tags": [
          "系统用户相关接口"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "",
            "required": false,
            "example": "",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": false,
            "example": "123",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-ignore-properties": [],
                  "x-apifox-orders": []
                }
              }
            }
          }
        },
        "x-run-in-apifox'": "https://www.apifox.cn/web/project/1888020/apis/api-49052334-run"
      },
      "put": {
        "summary": "修改系统用户(非密码字段)",
        "x-apifox-folder": "系统用户相关接口",
        "x-apifox-status": "testing",
        "deprecated": false,
        "description": "`SportsCompanyID` 和 `SportsCompanyName` 由 前端 填写，后端不会进行干预。\n`ModifyDatetime` 和 `CreateDatetime` 无论前端填写什么内容，后端都会干预。\n`Creator` 和 `Modifier` 若前端未填写，则后端会默认使用 `admin` 作为数据修改/新增者。\n",
        "tags": [
          "系统用户相关接口"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": false,
            "example": "123",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "x-apifox-overrides": {},
                "type": "object",
                "x-apifox-refs": {
                  "01GJZRT6ETYQH8WNCVE1J4KBA1": {
                    "$ref": "#/components/schemas/SysUser"
                  }
                },
                "x-apifox-orders": [
                  "01GJZRT6ETYQH8WNCVE1J4KBA1"
                ],
                "properties": {
                  "Id": {
                    "type": "string"
                  },
                  "LoginName": {
                    "type": "string"
                  },
                  "DisplayName": {
                    "type": "string"
                  },
                  "Cellphone": {
                    "type": "string"
                  },
                  "Email": {
                    "type": "string"
                  },
                  "Wechat": {
                    "type": "string"
                  },
                  "QQ": {
                    "type": "string"
                  },
                  "IsDisableLogin": {
                    "type": "boolean"
                  },
                  "CreateDatetime": {
                    "type": "null"
                  },
                  "ModifyDatetime": {
                    "type": "null"
                  },
                  "Creator": {
                    "type": "string"
                  },
                  "Modifier": {
                    "type": "string",
                    "nullable": true
                  },
                  "PasswdMD5": {
                    "type": "string"
                  },
                  "SportsCompanyId": {
                    "type": "string",
                    "maxLength": 36,
                    "format": "uuid",
                    "title": "体育公司唯一编号"
                  },
                  "SportsCompanyName": {
                    "type": "string",
                    "maxLength": 36,
                    "title": "体育公司名称"
                  },
                  "IDCardNumber": {
                    "type": "string",
                    "maxLength": 18,
                    "pattern": "[1-9]\\d{5}(18|19|20)\\d{2}((0[1-9])|(1[0-2]))(([0-2][1-9])|10|20|30|31)\\d{3}[0-9Xx]",
                    "title": "身份证号码"
                  }
                },
                "required": [
                  "Id",
                  "LoginName",
                  "DisplayName",
                  "Cellphone",
                  "Email",
                  "Wechat",
                  "QQ",
                  "IsDisableLogin",
                  "Modifier",
                  "SportsCompanyId",
                  "SportsCompanyName",
                  "IDCardNumber"
                ],
                "x-apifox-ignore-properties": [
                  "Id",
                  "LoginName",
                  "DisplayName",
                  "Cellphone",
                  "Email",
                  "Wechat",
                  "QQ",
                  "IsDisableLogin",
                  "CreateDatetime",
                  "ModifyDatetime",
                  "Creator",
                  "Modifier",
                  "PasswdMD5",
                  "SportsCompanyId",
                  "SportsCompanyName",
                  "IDCardNumber"
                ]
              },
              "example": {}
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-ignore-properties": [],
                  "x-apifox-orders": []
                }
              }
            }
          }
        },
        "x-run-in-apifox'": "https://www.apifox.cn/web/project/1888020/apis/api-51883211-run"
      }
    },
    "/login": {
      "post": {
        "summary": "系统用户登录",
        "x-apifox-folder": "系统用户相关接口",
        "x-apifox-status": "testing",
        "deprecated": false,
        "description": "",
        "tags": [
          "系统用户相关接口"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "title": "",
                "type": "object",
                "properties": {
                  "login_name": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "code": {
                    "type": "string"
                  }
                },
                "x-apifox-orders": [
                  "login_name",
                  "password",
                  "code"
                ],
                "required": [
                  "login_name",
                  "password",
                  "code"
                ],
                "x-apifox-ignore-properties": []
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-ignore-properties": [],
                  "x-apifox-orders": []
                }
              }
            }
          }
        },
        "x-run-in-apifox'": "https://www.apifox.cn/web/project/1888020/apis/api-48247332-run"
      }
    },
    "/athorizationinfo": {
      "get": {
        "summary": "获取登录信息",
        "x-apifox-folder": "系统用户相关接口",
        "x-apifox-status": "testing",
        "deprecated": false,
        "description": "",
        "tags": [
          "系统用户相关接口"
        ],
        "parameters": [
          {
            "name": "field",
            "in": "query",
            "description": "支持 acl,login_name,token,sysuser,display_name 的获取",
            "required": true,
            "example": "acl",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": true,
            "example": "90bb3fc6aab0fd170675515d6226ea1af6358c4e",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-ignore-properties": [],
                  "x-apifox-orders": []
                }
              }
            }
          }
        },
        "x-run-in-apifox'": "https://www.apifox.cn/web/project/1888020/apis/api-48287964-run"
      }
    },
...
    "/role": {
      "post": {
        "summary": "新增角色",
        "x-apifox-folder": "系统管理相关接口/权限角色相关接口",
        "x-apifox-status": "testing",
        "deprecated": false,
        "description": "`ID` 创建时会被后端功能干预，因此可以填写空字符串。\n`ModifyDateTime` 和 `CreateDateTime` 后台会进行干预。\n`Creator` 和 `Modifier` 前端不填写的话，后端会自动填写为 \"admin\"",
        "tags": [
          "系统管理相关接口/权限角色相关接口"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SysRole",
                "x-apifox-overrides": {}
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-ignore-properties": [],
                  "x-apifox-orders": []
                }
              }
            }
          }
        },
        "x-run-in-apifox'": "https://www.apifox.cn/web/project/1888020/apis/api-49518126-run"
      },
      "get": {
        "summary": "获取系统角色信息",
        "x-apifox-folder": "系统管理相关接口/权限角色相关接口",
        "x-apifox-status": "testing",
        "deprecated": false,
        "description": "",
        "tags": [
          "系统管理相关接口/权限角色相关接口"
        ],
        "parameters": [
          {
            "name": "role_id",
            "in": "query",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": true,
            "example": "test",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-ignore-properties": [],
                  "x-apifox-orders": []
                }
              }
            }
          }
        },
        "x-run-in-apifox'": "https://www.apifox.cn/web/project/1888020/apis/api-55991085-run"
      },
      "put": {
        "summary": "修改角色信息",
        "x-apifox-folder": "系统管理相关接口/权限角色相关接口",
        "x-apifox-status": "developing",
        "deprecated": false,
        "description": "",
        "tags": [
          "系统管理相关接口/权限角色相关接口"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": true,
            "example": "test",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SysRole",
                "x-apifox-overrides": {}
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-ignore-properties": [],
                  "x-apifox-orders": []
                }
              }
            }
          }
        },
        "x-run-in-apifox'": "https://www.apifox.cn/web/project/1888020/apis/api-55992143-run"
      }
    },
    "/sysfuncpage/list": {
      "get": {
        "summary": "获取页面及功能列表",
        "x-apifox-folder": "系统管理相关接口/权限角色相关接口",
        "x-apifox-status": "testing",
        "deprecated": false,
        "description": "",
        "tags": [
          "系统管理相关接口/权限角色相关接口"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {},
                "x-apifox-ignore-properties": [],
                "x-apifox-orders": []
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SysFuncPage1",
                  "x-apifox-overrides": {}
                }
              }
            }
          }
        },
        "x-run-in-apifox'": "https://www.apifox.cn/web/project/1888020/apis/api-49541498-run"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "SysUser": {
        "type": "object",
        "properties": {
          "Id": {
            "type": "string"
          },
          "LoginName": {
            "type": "string"
          },
          "DisplayName": {
            "type": "string"
          },
          "Cellphone": {
            "type": "string"
          },
          "Email": {
            "type": "string"
          },
          "Wechat": {
            "type": "string"
          },
          "QQ": {
            "type": "string"
          },
          "IsDisableLogin": {
            "type": "boolean"
          },
          "CreateDatetime": {
            "type": "null"
          },
          "ModifyDatetime": {
            "type": "null"
          },
          "Creator": {
            "type": "string"
          },
          "Modifier": {
            "type": "string"
          },
          "PasswdMD5": {
            "type": "string"
          },
          "SportsCompanyId": {
            "type": "string",
            "maxLength": 36,
            "format": "uuid",
            "title": "体育公司唯一编号"
          },
          "SportsCompanyName": {
            "type": "string",
            "maxLength": 36,
            "title": "体育公司名称"
          },
          "IDCardNumber": {
            "type": "string",
            "maxLength": 18,
            "pattern": "[1-9]\\d{5}(18|19|20)\\d{2}((0[1-9])|(1[0-2]))(([0-2][1-9])|10|20|30|31)\\d{3}[0-9Xx]",
            "title": "身份证号码"
          }
        },
        "required": [
          "Id",
          "LoginName",
          "DisplayName",
          "Cellphone",
          "Email",
          "Wechat",
          "QQ",
          "IsDisableLogin",
          "Creator",
          "Modifier",
          "SportsCompanyId",
          "SportsCompanyName",
          "IDCardNumber"
        ],
        "x-apifox-orders": [
          "Id",
          "LoginName",
          "DisplayName",
          "Cellphone",
          "Email",
          "Wechat",
          "QQ",
          "IsDisableLogin",
          "CreateDatetime",
          "ModifyDatetime",
          "Creator",
          "Modifier",
          "PasswdMD5",
          "SportsCompanyId",
          "SportsCompanyName",
          "IDCardNumber"
        ],
        "title": "系统用户",
        "x-apifox-ignore-properties": [],
        "x-apifox-folder": ""
      },
...
    }
  },
  "servers": []
}